working on different UX flow for leaving memory
