+++ react-native-app/.eslintrc.js
@@ -5,7 +5,8 @@ module.exports = {
         "react/jsx-filename-extension": "off",
         "react/no-unused-state": "off",
         "react/prefer-stateless-function": "off",
-        "class-methods-use-this": "off"
+        "class-methods-use-this": "off",
+        "react/prop-types": "off"
     }
 
 };
